Fix remaining lint issues on touched lines
- Suppress pre-existing staticcheck deprecation warnings (SA1019)
- Refactor go-require: move assertions out of goroutine

+++ tests/schedule_test.go
@@ -302,7 +302,7 @@ func (s *scheduleFunctionalSuiteBase) TestBasics() {
 	s.GreaterOrEqual(len(describeResp.Info.RecentActions), 2)
 	action0 := describeResp.Info.RecentActions[0]
 	s.WithinRange(action0.ScheduleTime.AsTime(), createTime, time.Now())
-	s.Equal(0, action0.ScheduleTime.AsTime().UnixNano()%int64(5*time.Second))
+	s.Equal(int64(0), action0.ScheduleTime.AsTime().UnixNano()%int64(5*time.Second))
 	s.DurationNear(action0.ActualTime.AsTime().Sub(action0.ScheduleTime.AsTime()), 0, 3*time.Second)
 
 	// validate list response
@@ -350,7 +350,7 @@ func (s *scheduleFunctionalSuiteBase) TestBasics() {
 	var ex0StartTime time.Time
 	s.NoError(payload.Decode(ex0.SearchAttributes.IndexedFields[sadefs.TemporalScheduledStartTime], &ex0StartTime))
 	s.WithinRange(ex0StartTime, createTime, time.Now())
-	s.Equal(0, ex0StartTime.UnixNano()%int64(5*time.Second))
+	s.Equal(int64(0), ex0StartTime.UnixNano()%int64(5*time.Second))
 
 	// list schedules with search attribute filter
 
@@ -415,7 +415,7 @@ func (s *scheduleFunctionalSuiteBase) TestBasics() {
 
 	s.DurationNear(describeResp.Info.UpdateTime.AsTime().Sub(updateTime), 0, 3*time.Second)
 	lastAction := describeResp.Info.RecentActions[len(describeResp.Info.RecentActions)-1]
-	s.Equal(1000000000, lastAction.ScheduleTime.AsTime().UnixNano()%int64(5*time.Second), lastAction.ScheduleTime.AsTime().UnixNano())
+	s.Equal(int64(1000000000), lastAction.ScheduleTime.AsTime().UnixNano()%int64(5*time.Second), lastAction.ScheduleTime.AsTime().UnixNano())
 
 	// update schedule and search attributes
